Added <= <, >, >= but not == to XMoney.  Operations convert to the same currency as the left operand.  == is not defined because that would either make the class unhashable, or make hashing an expensive operation.
